threagile-builder/requirements.txt
# threagile-builder/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
apiflask==2.2.1
black==24.10.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-exporter-prometheus==0.48b0
prefect==2.20.10
prometheus_client==0.15.0
python-dotenv==1.0.1